requirements.txt
pyrogram
tgcrypto
python-dotenv
motor
pymongo